# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BE Nhựa API",
        "description": "API quản lý sản phẩm và biến thể sản phẩm",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Current server"
        }
    ],
    "paths": {
...
        "/api/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Danh sách sản phẩm",
                "operationId": "16151fd150adb5c53ccb365b18872b88",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công"
                    },
                    "401": {
                        "description": "Chưa xác thực"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Tạo sản phẩm kèm ảnh",
                "operationId": "39e039020db35ef73fb90039e9a2e0fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Đã tạo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Chi tiết sản phẩm",
                "operationId": "e8be9e42b1e86fa64b22150625f74dbd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tồn tại"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Cập nhật sản phẩm kèm ảnh",
                "description": "Gửi multipart/form-data bằng POST và đặt _method=PUT.",
                "operationId": "5538a3965ce7bdd5a876cc61568299cd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đã cập nhật",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/ProductResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Không tồn tại"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Xóa sản phẩm",
                "operationId": "4a840887f9748a17d79fe0af9ab29131",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Đã xóa"
                    },
                    "404": {
                        "description": "Không tồn tại"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreProductRequest": {
                "required": [
                    "product_name"
                ],
                "properties": {
                    "product_name": {
                        "type": "string",
                        "example": "Áo thun basic",
                        "maxLength": 255
                    },
                    "sku": {
                        "type": "string",
                        "example": "AT-001",
                        "nullable": true,
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean",
                        "nullable": true,
                        "default": true
                    },
                    "is_featured": {
                        "type": "boolean",
                        "nullable": true,
                        "default": false
                    },
                    "variant_groups": {
                        "type": "array",
                        "items": {
                            "required": [
                                "variant_group_id"
                            ],
                            "properties": {
                                "variant_group_id": {
                                    "type": "integer"
                                },
                                "is_required": {
                                    "type": "boolean"
                                },
                                "sort_order": {
                                    "type": "integer",
                                    "minimum": 0
                                }
                            },
                            "type": "object"
                        }
                    },
                    "images[]": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        },
                        "maxItems": 10
                    }
                },
                "type": "object"
            },
            "UpdateProductRequest": {
                "properties": {
                    "_method": {
                        "type": "string",
                        "example": "PUT"
                    },
                    "product_name": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "sku": {
                        "type": "string",
                        "nullable": true,
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "is_featured": {
                        "type": "boolean"
                    },
                    "variant_groups": {
                        "type": "array",
                        "items": {
                            "required": [
                                "variant_group_id"
                            ],
                            "properties": {
                                "variant_group_id": {
                                    "type": "integer"
                                },
                                "is_required": {
                                    "type": "boolean"
                                },
                                "sort_order": {
                                    "type": "integer",
                                    "minimum": 0
                                }
                            },
                            "type": "object"
                        }
                    },
                    "images[]": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "binary"
                        },
                        "maxItems": 10
                    }
                },
                "type": "object"
            },
...
            "FileResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "file_name": {
                        "type": "string",
                        "nullable": true
                    },
                    "disk": {
                        "type": "string",
                        "example": "public"
                    },
                    "path": {
                        "type": "string",
                        "example": "products/1/image.jpg"
                    },
                    "mime_type": {
                        "type": "string",
                        "example": "image/jpeg"
                    },
                    "size": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "type": {
                        "type": "string",
                        "example": "image",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_name": {
                        "type": "string"
                    },
                    "sku": {
                        "type": "string",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "is_featured": {
                        "type": "boolean"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FileResource"
                        }
                    },
                    "variant_groups": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VariantGroupResource"
                        }
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ProductVariantResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ProductVariantResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "product_id": {
                        "type": "integer"
                    },
                    "sku": {
                        "type": "string"
                    },
                    "price": {
                        "type": "string",
                        "example": "150000.00",
                        "nullable": true
                    },
                    "stock": {
                        "type": "integer"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "images": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FileResource"
                        }
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VariantOptionResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "VariantGroupResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "group_code": {
                        "type": "string"
                    },
                    "group_name": {
                        "type": "string"
                    },
                    "sort_order": {
                        "description": "Chỉ có khi group thuộc product",
                        "type": "integer",
                        "nullable": true
                    },
                    "is_required": {
                        "description": "Chỉ có khi group thuộc product",
                        "type": "boolean",
                        "nullable": true
                    },
                    "options": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/VariantOptionResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "VariantOptionResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "variant_group_id": {
                        "type": "integer"
                    },
                    "option_code": {
                        "type": "string"
                    },
                    "option_name": {
                        "type": "string"
                    },
                    "sort_order": {
                        "type": "integer"
                    },
                    "is_active": {
                        "type": "boolean"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "Bearer token",
                "scheme": "bearer"
            }
        }
    },
...
}